+++ CHANGES.md
@@ -5,6 +5,7 @@
 - Double clicking the headers in the Week/All Events views (for a week/month) will now activate the expand button.
 - Improved the "No events available" default text display in the Week view.
 - An error log is now shown in the developer console when an input type is not supported by the browser.
+- Added public function updateEvents(), which will update an array of existing events.
 - 
 
 ### Version 0.9.4: